requirements.txt
python-dotenv
openai
aiogram
django